re-organised code in theme folder
