+++ src/components/Example/styles.ts
@@ -3,6 +3,7 @@ import { makeStyles, Theme } from '@material-ui/core/styles';
 const useStyles = makeStyles((theme: Theme) => ({
   button: {
     backgroundColor: theme.palette.highlight,
+    color: theme.palette.text.secondary,
   },
 }));
